# backend/Tempswagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trilho API",
    "version": "v1"
  },
  "paths": {
    "/api/admin/auth": {
      "post": {
        "tags": [
          "AdminEndpoints"
        ],
        "operationId": "AdminLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "AdminEndpoints"
        ],
        "operationId": "AdminGetUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}/vip": {
      "patch": {
        "tags": [
          "AdminEndpoints"
        ],
        "operationId": "AdminPatchVip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchVipDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/lines/{code}/status": {
      "get": {
        "tags": [
          "LineEndpoints"
        ],
        "operationId": "GetLineStatus",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lines/{code}/vehicles": {
      "get": {
        "tags": [
          "LineEndpoints"
        ],
        "operationId": "GetLineVehicles",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/register-token": {
      "post": {
        "tags": [
          "NotificationEndpoints"
        ],
        "operationId": "RegisterFcmToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/token/{token}": {
      "delete": {
        "tags": [
          "NotificationEndpoints"
        ],
        "operationId": "UnregisterFcmToken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/notifications/subscribe/{lineCode}": {
      "post": {
        "tags": [
          "NotificationEndpoints"
        ],
        "operationId": "SubscribeLineNotifications",
        "parameters": [
          {
            "name": "lineCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "NotificationEndpoints"
        ],
        "operationId": "UnsubscribeLineNotifications",
        "parameters": [
          {
            "name": "lineCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PatchVipDto": {
        "type": "object",
        "properties": {
          "isVip": {
            "type": "boolean"
          },
          "vipEmail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterTokenRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}